+++ pytrading212/__init__.py
@@ -1,3 +1,3 @@
 from .order import *
 from .position import Position
-from .trading212 import CFD, Equity, Trading212, Mode
+from .trading212 import CFD, Trading212, Mode